# docs/swagger/swagger.yaml
components:
  examples: {}
  headers: {}
  parameters: {}
  requestBodies: {}
  responses: {}
  schemas: { Secret: { properties: { value: {} }, type: object } }
  securitySchemes: {}
info:
  title: '@express/vault'
  version: 0.0.1
  description: 'express.js - vault'
  license: { name: 'GNU GPL 3.0' }
  contact: { name: 'hieudoanm ', email: hieumdoan@gmail.com }
openapi: 3.0.0
paths:
  /health:
    {
      get:
        {
          operationId: GetHealth,
          responses:
            {
              'Health Check':
                {
                  description: 'Health Check',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              properties: { status: { type: string, enum: [OK], nullable: false } },
                              required: [status],
                              type: object,
                            },
                        },
                    },
                },
            },
          tags: [Health],
          security: [],
          parameters: [],
        },
    }
  '/secrets/{key}':
    {
      get:
        {
          operationId: GetSecret,
          responses:
            {
              '200':
                { description: Ok, content: { application/json: { schema: { $ref: '#/components/schemas/Secret' } } } },
            },
          tags: [Secrets],
          security: [],
          parameters: [{ in: path, name: key, required: true, schema: { type: string } }],
        },
      post:
        {
          operationId: SetSecret,
          responses: { '200': { description: Ok, content: { application/json: { schema: { type: boolean } } } } },
          tags: [Secrets],
          security: [],
          parameters: [{ in: path, name: key, required: true, schema: { type: string } }],
          requestBody:
            { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Secret' } } } },
        },
    }
servers:
  - { url: 'http://localhost:8080', description: Local }